# api-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IdentyClaw Federated Peer API (RODiT)",
    "version": "2025.10.01",
...
    "termsOfService": "https://identyclaw.com",
    "contact": {
      "name": "Discernible Inc.",
      "url": "https://identyclaw.com",
      "email": "support@rodit-tech.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://identyclaw.com"
    },
    "x-api-id": "identyclaw-federated-peer-v1"
  },
  "servers": [
    {
      "url": "https://slc.discernible.io:8443",
      "description": "Example peer (main) — replace with your federated peer base URL"
    },
    {
      "url": "https://slc.dihola.io:8443",
      "description": "Example peer (development)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Node.js (direct, no nginx TLS)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "RODiT-JWT",
        "description": "Peer-minted JWT from POST /api/login on **this** apiEndpoint (`jwt_token` in body). Not a cookie session. Not interchangeable with IdentyClaw home JWTs or other peers' tokens."
      }
    },
...
  },
...
  "paths": {
...
    "/api/sessions/list_all": {
      "get": {
        "summary": "List all sessions",
        "description": "Get information about all active sessions (admin only)",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions retrieved successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      }
    },
    "/api/sessions/revoke": {
      "post": {
        "summary": "Revoke specific session",
        "description": "Force terminate a specific session (admin only)",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string",
                    "description": "ID of session to terminate"
                  },
                  "reason": {
                    "type": "string",
                    "description": "Reason for termination"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session terminated successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission denied"
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/api/sessions/cleanup": {
      "post": {
        "summary": "Cleanup expired sessions",
        "description": "Run manual cleanup of expired sessions (admin only)",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cleanup completed successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      }
    },
...
  },
...
}